replase def get_port_and_address_for_use in common/utils.py
